Conflicts:
	conf/logging.json
	conf/logging/client.json
	conf/properties/run
	src/main/java/com/emc/mongoose/base/load/client/impl/BasicLoadClient.java
	src/main/java/com/emc/mongoose/base/load/impl/LoadExecutorBase.java
	src/main/java/com/emc/mongoose/base/load/impl/StorageNodeExecutorBase.java
	src/main/java/com/emc/mongoose/base/load/type/CreateLoadBase.java
	src/main/java/com/emc/mongoose/web/api/impl/WSRequestConfigBase.java
	src/main/java/com/emc/mongoose/web/api/impl/provider/atmos/RequestConfig.java
	src/main/java/com/emc/mongoose/web/api/impl/provider/s3/RequestConfig.java
	src/main/java/com/emc/mongoose/web/api/impl/provider/swift/RequestConfig.java

+++ src/main/java/com/emc/mongoose/web/load/impl/WSLoadHelper.java
@@ -37,7 +37,7 @@ public class WSLoadHelper {
 		connMgr.setDefaultConnectionConfig(
 			ConnectionConfig
 				.custom()
-				.setBufferSize((int) dataPageSize)
+				.setBufferSize(dataPageSize)
 				.build()
 		);
 		// set shared headers to client builder
